+++ src/eca/features/commands.clj
@@ -57,20 +57,21 @@
   (->> (get config :commands)
        (map
         (fn [{:keys [path]}]
-          (if (fs/absolute? path)
-            (when (fs/exists? path)
-              {:name (normalize-command-name path)
-               :path path
-               :type :user-config
-               :content (slurp path)})
-            (keep (fn [{:keys [uri]}]
-                    (let [f (fs/file (shared/uri->filename uri) path)]
-                      (when (fs/exists? f)
-                        {:name (normalize-command-name f)
-                         :path (str (fs/canonicalize f))
-                         :type :user-config
-                         :content (slurp f)})))
-                  roots))))
+          (let [path (str (fs/expand-home path))]
+            (if (fs/absolute? path)
+              (when (fs/exists? path)
+                {:name (normalize-command-name path)
+                 :path path
+                 :type :user-config
+                 :content (slurp path)})
+              (keep (fn [{:keys [uri]}]
+                      (let [f (fs/file (shared/uri->filename uri) path)]
+                        (when (fs/exists? f)
+                          {:name (normalize-command-name f)
+                           :path (str (fs/canonicalize f))
+                           :type :user-config
+                           :content (slurp f)})))
+                    roots)))))
        (flatten)
        (remove nil?)))